bugfix : getStatError -> getRelStatError
getRate always expects a relative error since it does rate*(1+epsilon)
However, the stat err from the histogram was fed as an absolute error.
This is still not the correct thing to do when adding in quadrature
all the different error sources, but it's good enough for a test to
just check the new errors (which are all included in the error bar of
the matrix, i.e. adding in quadrature only one source).