uniquefy print_available_regions.get_histonames
This will only have an effect when there are multiple versions of a
given histogram (for example when you update the file and hack one of
them).  Without this change, compare_matrices.py will not work
correctly.  Note that by default root will pick the most recent one
when there are two keys with the same name, so this should be always
safe.