add razor systematics
These systematic uncertainties are based on what we had for the WH
weighted average.

Because we used a 2d parametrization, and for the razor we're now
using a 1d parametrization, I had to change a few things in addition
to just adding the necessary histograms.  In particular, this change
breaks the 2d printout that was implemented in
test_printSystematics2d.